Power rates are now doubles to support accuracy in the sub-mW range.

+++ org.qualitune.jouleunit/src/org/qualitune/jouleunit/AbstractJouleProfiler.java
@@ -447,7 +447,7 @@ public abstract class AbstractJouleProfiler extends AbstractProfilerNTimes {
 	protected PowerRate getPowerRateValueProbablyWithNoiseEstimation() {
 
 		PowerRate rateValue = getPowerRateValue();
-		long powerRate = rateValue.getPowerRate();
+		double powerRate = rateValue.getPowerRate();
 
 		if (isProbeEffectCosiderationEnabled || isBaseNoiseCosiderationEnabled) {
